requirements.txt
click
flask
python-socketio
requests